+++ players/urls.py
@@ -7,5 +7,7 @@ urlpatterns = [
     path('Players/<uuid:pk>', PlayersDetail.as_view()),
     path('Players/<uuid:pk>/Vote', PlayerVote.as_view()),
     path('Players/<uuid:pk>/Votes', PlayerVotes.as_view()),
-    path('Players/<uuid:subjpk>/Vote/<uuid:objpk>', PlayerVoteFor.as_view())
+    path('Players/<uuid:subjpk>/Vote/<uuid:objpk>', PlayerVoteFor.as_view()),
+    path('AllVoted', AllVoted.as_view()),
+    path('Votes', ClearVotes.as_view())
 ]